Added `method` parameter to stubRequest
The jsdoc comment mentioned the posibility to specify what method to
stub, but the actual code didn't support it. Now it does.

To match the documentation and other stub functions the `method`
parameter is the first parameter. To keep the code backwards
compatible this parameter needed to be optional. That unfortunately
made the actual implementation a litte bit messy.

Fixes #5